+++ src/com/kovalenych/media/VideoResponse.java
@@ -12,7 +12,7 @@ public class VideoResponse {
     public long maxId;
 
     @SerializedName("since_id")
-    public int sinceId;
+    public long sinceId;
 
     public String query;
